add index_ops.h
